Improve hashing & topics
Topics now support math & predicates (similar to Head in some ways, these let you have a bool made out of underlying bools such as "is a set full"...)
Billboard supports offset to make it easier. I've standardized on:
The "floor" is at +.5f, so that the collider plane is 0f. This means sprites need an offset of +.5f so their feet are on the ground.