Add panning code, make external builds work again
It looks like Val is poison: something about that forces unity to figure out serialization for abstract ValueTopic<T> and get hung up on the [SerializeField] T defaultValue & value fields.
I thought it might be the static readonly Eq, but removing it didn't fix things (or harm them, so I left it out).
But removing the Val class let it build again so...?!