Fix converter & serialization
The code was missing the case A->A, which (foolishly?) isn't considered a cast. The new version catches IsAssignable and then uses a switch to force the cast; I'm not convinced this is efficient but it's goodish enough-ish.
It now *requires* that if there is a fooplicit operator, that it support `default(TIn)` without throwing (and since the actual use of the code is symmetric, that means the reverse as well).

this makes it so that you can view object references out of topics