Expanded attribute; revamp cloning & singleton assets, neither of which was Prod Ready.
This makes it possible to inline edit the scriptable objects that are created.

Since we have singletons now, use that for cloning (instead of the careful editor linkages which were NOT getting maintained...).
Modify scriptable object to work safely the way I intend it to (like: automatically following typenames when you create an asset from the script...).