Fix custom drawers: drive events right
We don't need (???) to call serializedObject.Update/setValues; my guess is that that's the old/editor way of doing it, and that drawers can't call that, and nobody !@##ing documented it.

Instead, special case the label (so that Subtype<T> gets labeled, and [Subtype] does NOT since it expects to draw on top of an expanded object.

Add singleton TopicDebugging which can do "stuff" to stitch together topics, not clear what exactly atm.

Add read only drawer.

Kill "asset bundle"; I can resurrect it but I think it's _literally_ rebuilding asset bundle and that feels bad to embark on.